requirements.txt
flask 
flask_sqlalchemy
flask_wtf
selenium
requests
marshmallow
werkzeug
PyJWT
